+++ systemv/cupsctl.c
@@ -91,6 +91,9 @@ main(int  argc,				/* I - Number of command-line args */
         if (!strcmp(argv[i], "--debug-logging"))
 	  num_settings = cupsAddOption(CUPS_SERVER_DEBUG_LOGGING, "1",
 	                               num_settings, &settings);
+        else if (!strcmp(argv[i], "--verbose-debug-logging"))
+	  num_settings = cupsAddOption(CUPS_SERVER_DEBUG_LOGGING, "2",
+	                               num_settings, &settings);
         else if (!strcmp(argv[i], "--no-debug-logging"))
 	  num_settings = cupsAddOption(CUPS_SERVER_DEBUG_LOGGING, "0",
 	                               num_settings, &settings);
@@ -226,14 +229,14 @@ usage(const char *opt)			/* I - Option character/string */
 
   _cupsLangPuts(stdout, _("Usage: cupsctl [options] [param=value ... paramN=valueN]"));
   _cupsLangPuts(stdout, _("Options:"));
-  _cupsLangPuts(stdout, _("-E                      Encrypt the connection to the server"));
-  _cupsLangPuts(stdout, _("-h server[:port]        Connect to the named server and port"));
-  _cupsLangPuts(stdout, _("-U username             Specify username to use for authentication"));
-  _cupsLangPuts(stdout, _("--[no-]debug-logging    Turn debug logging on/off"));
-  _cupsLangPuts(stdout, _("--[no-]remote-admin     Turn remote administration on/off"));
-  _cupsLangPuts(stdout, _("--[no-]remote-any       Allow/prevent access from the Internet"));
-  _cupsLangPuts(stdout, _("--[no-]share-printers   Turn printer sharing on/off"));
-  _cupsLangPuts(stdout, _("--[no-]user-cancel-any  Allow/prevent users to cancel any job"));
+  _cupsLangPuts(stdout, _("-E                               Encrypt the connection to the server"));
+  _cupsLangPuts(stdout, _("-h server[:port]                 Connect to the named server and port"));
+  _cupsLangPuts(stdout, _("-U username                      Specify username to use for authentication"));
+  _cupsLangPuts(stdout, _("--[no-|verbose-]debug-logging    Turn debug logging on/off, --verbose prefix enables extensive logging"));
+  _cupsLangPuts(stdout, _("--[no-]remote-admin              Turn remote administration on/off"));
+  _cupsLangPuts(stdout, _("--[no-]remote-any                Allow/prevent access from the Internet"));
+  _cupsLangPuts(stdout, _("--[no-]share-printers            Turn printer sharing on/off"));
+  _cupsLangPuts(stdout, _("--[no-]user-cancel-any           Allow/prevent users to cancel any job"));
 
   exit(1);
 }